[feat] fix openen fragemnt

+++ app/src/main/java/com/klimaatmobiel/domain/Group.kt
@@ -5,7 +5,7 @@ import com.squareup.moshi.Json
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
-class Group(val groupId: Long, val groupName: String, val projectId: Double, val project : Project,
+class Group(val groupId: Long, val groupName: String, val projectId: Long, val project : Project,
             var order : Order, val uniqueGroupCode: String) : Parcelable {
